refactor(ctl-api): extract app-config graph and install validation into internal/pkg
Move component enablement graph resolution into internal/pkg/appconfiggraph
(split into a pure structural Graph plus a ComponentEnablementResolver) and
relocate installvalidate from internal/app to internal/pkg. Update install
component read, input-update, and workflow callers to the new packages. No
behavior change.
