+++ services/ctl-api/internal/pkg/installvalidate/diagnostics.go
